refactoring delete useless packages
