--- go/trip/service_test.go
@@ -1,9 +1,0 @@
-package trip
-
-import (
-	"testing"
-)
-
-func TestGetGetTripByUser(t *testing.T) {
-
-}
