--- go/trip/trip.go
@@ -1,4 +1,0 @@
-package trip
-
-type Trip struct {
-}
