--- go/user/session.go
@@ -1,19 +1,0 @@
-package user
-
-import "errors"
-
-type Session struct {
-}
-
-var session *Session
-
-func CurrentSession() *Session {
-	if session == nil {
-		session = &Session{}
-	}
-	return session
-}
-
-func (this *Session) GetLoggedUser() (*User, error) {
-	return nil, errors.New("Session.GetLoggedUser() should not be called in an unit test")
-}
