--- go/user/user.go
@@ -1,14 +1,0 @@
-package user
-
-type User struct {
-}
-
-func NewUser() *User {
-	return &User{}
-}
-
-func (this *User) Friends() ([]User, error) {
-	var friends []User
-
-	return friends, nil
-}
